+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 
 * DNS
   * Fixed issue in the `akamai_dns_record` resource where updates to records with multi-value targets were not handled correctly ([I#767](https://github.com/akamai/terraform-provider-akamai/issues/767)).
+  * Fixed issue in the `akamai_dns_record` resource where updates to CAA records could cause unnecessary diffs due to normalization of the data.